+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * add additional external chapters when creating a CLDF dataset
 * `topics_index.csv` will be added as a TopicTable
 * `entity_refs.yaml` allows to add section references to entries in the database
+* `full` argument for citations
 
 ### Fixed
 * LaTeX examples